commit by zhoulinhui 新建了四个模块

+++ malldemo/oomall/adModule/src/main/java/xmu/oomall/ad/controller/AdController.java
@@ -5,8 +5,11 @@ import org.springframework.util.StringUtils;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import xmu.oomall.ad.domain.Ad;
+import xmu.oomall.ad.domain.Log;
 import xmu.oomall.ad.service.impl.AdService;
-import xmu.oomall.util.ResponseUtil;
+import xmu.oomall.ad.util.ResponseUtil;
+
+import javax.servlet.http.HttpServletRequest;
 
 
 @RestController
@@ -37,11 +40,22 @@ public class AdController {
     * @Date: 2019/12/10
     */
     @GetMapping("/admins/ads")
-    public Object adminFindAdList(@RequestParam(defaultValue = "1") Integer page,
+    public Object adminFindAdList(HttpServletRequest request,
+                                  @RequestParam(defaultValue = "1") Integer page,
                                   @RequestParam(defaultValue = "10") Integer limit,
                                   @RequestParam String adTitle,
                                   @RequestParam String adContent){
-
+        String id= request.getHeader("id");
+        if (id==null){
+            return ResponseUtil.unlogin();
+        }
+        Log log=new Log();
+        log.setAdminId(Integer.valueOf(id));
+        log.setIp(request.getRemoteAddr());
+        log.setType(0);
+        log.setStatusCode(1);
+        log.setActions("查询列表");
+        adService.log(log);
         Ad ad=new Ad();
         ad.setName(adTitle);
         ad.setContent(adContent);
@@ -70,11 +84,25 @@ public class AdController {
     * @Date: 2019/12/5
     */
     @GetMapping("/ads/{id}")
-    public Object adminFindAdById(@PathVariable Integer id){
+    public Object adminFindAdById(HttpServletRequest request,@PathVariable Integer id){
+        String adminid= request.getHeader("id");
+        if (adminid==null){
+            return ResponseUtil.unlogin();
+        }
+        Log log=new Log();
+        log.setAdminId(Integer.valueOf(adminid));
+        log.setIp(request.getRemoteAddr());
+        log.setType(0);
+
+        log.setActions("查询详情");
+        adService.log(log);
         Ad adById = adService.findAdById(id);
         if (adById==null){
+            log.setStatusCode(0);
+            adService.log(log);
             return ResponseUtil.fail(690,"广告不存在");
         }
+        adService.log(log);
         return  ResponseUtil.ok(adById);
     }
 
@@ -88,7 +116,16 @@ public class AdController {
     * @Date: 2019/12/7
     */
     @PutMapping("/ads/{id}")
-    public Object adminUpdateAd(@PathVariable Integer id,@RequestBody Ad newAd) {
+    public Object adminUpdateAd(HttpServletRequest request,@PathVariable Integer id,@RequestBody Ad newAd) {
+        String adminid= request.getHeader("id");
+        if (adminid==null){
+            return ResponseUtil.unlogin();
+        }
+        Log log=new Log();
+        log.setAdminId(Integer.valueOf(adminid));
+        log.setIp(request.getRemoteAddr());
+        log.setType(0);
+        log.setActions("修改");
         Object error=validate(newAd);
         if (error != null) {
             return error;
@@ -96,8 +133,11 @@ public class AdController {
         newAd.setId(id);
         if(adService.updateAdById(newAd)==0)
         {
+            log.setStatusCode(0);
+            adService.log(log);
             return ResponseUtil.fail(691,"广告更新失败");
         }
+        adService.log(log);
         return ResponseUtil.ok(newAd);
     }
 
@@ -110,10 +150,22 @@ public class AdController {
     * @Date: 2019/12/7
     */
     @DeleteMapping("/ads/{id}")
-    public Object adminDeleteAd(@PathVariable Integer id){
+    public Object adminDeleteAd(HttpServletRequest request,@PathVariable Integer id){
+        String adminid= request.getHeader("id");
+        if (adminid==null){
+            return ResponseUtil.unlogin();
+        }
+        Log log=new Log();
+        log.setAdminId(Integer.valueOf(adminid));
+        log.setIp(request.getRemoteAddr());
+        log.setType(0);
+        log.setActions("删除");
         if (id == null) {
+            log.setStatusCode(0);
+            adService.log(log);
             return ResponseUtil.badArgument();
         }
+        adService.log(log);
         adService.deleteAdbyId(id);
         return ResponseUtil.ok();
     }
@@ -126,15 +178,26 @@ public class AdController {
     * @Date: 2019/12/5
     */
     @PostMapping("/ads")
-    public Object adminCreateAd(@RequestBody Ad ad) {
+    public Object adminCreateAd(HttpServletRequest request,@RequestBody Ad ad) {
+        String adminid= request.getHeader("id");
+        if (adminid==null){
+            return ResponseUtil.unlogin();
+        }
+        Log log=new Log();
+        log.setAdminId(Integer.valueOf(adminid));
+        log.setIp(request.getRemoteAddr());
+        log.setType(0);
+        log.setActions("新增");
       Object error = validate(ad);
         if (error != null) {
            return error;
         }
         if (adService.addAds(ad)!=1){
+            log.setStatusCode(0);
+            adService.log(log);
             return ResponseUtil.fail(691,"广告操作失败");
         }
-
+        adService.log(log);
         return ResponseUtil.ok(ad);
     }
